Make it build and install on Ubuntu 2020.04
nawk and oawk are currently desactivated. They didn't build correctly with
either heirloom yacc and lex, and Ubuntu's bison and flex. There were
crashes with heirloom yacc and lex so I wonder if they build correctly.
I've no time to investigate the issue now so I leave it as is.